+++ src/GSF.CarbonAware/src/GSF.CarbonAware.csproj
@@ -14,7 +14,7 @@
 
     <ItemGroup>
       <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-        <_Parameter1>GSF.CarbonIntensity.Tests</_Parameter1>
+        <_Parameter1>GSF.CarbonAware.Tests</_Parameter1>
       </AssemblyAttribute>
       <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
         <_Parameter1>DynamicProxyGenAssembly2</_Parameter1>
