+++ src/GSF.CarbonAware/src/Models/EmissionsData.cs
@@ -1,0 +1,18 @@
+namespace GSF.CarbonAware.Models;
+
+public record EmissionsData
+{
+    public string? Location { get; init; }
+    public DateTimeOffset Time { get; init; }
+    public double Rating { get; init; }
+    public TimeSpan Duration { get; set; }
+
+    public static implicit operator EmissionsData(global::CarbonAware.Model.EmissionsData emissionsData) {
+        return new EmissionsData {
+            Duration = emissionsData.Duration,
+            Location = emissionsData.Location,
+            Rating = emissionsData.Rating,
+            Time = emissionsData.Time
+        };
+    }
+}
